Add general-use functions
Functions to be used throughout the program. Memcpy copies a large
chunck of data from one location to another, Strcpy does a similar thing
for null-terminated strings, copying up to, but not including the null
char. WaitVBlank is currently a simple function that loops til a vblank
occurs. In future, this should be adapted to halt instead, in order to
reduce the gameboys power usage.