Create basic game structure
At the moment, the game handles loading a font into memory, writing some
test text to the screen and initialising the cursor.

The current cursor position is stored a number between 1 and CURSOR_MAX,
where each possible location is a different index. The location in VRAM
of each cursor is stored in a lookup table, CURSOR_TABLE, which is
filled during assembly using a macro. This means that getting a cursor
position is as simple as loading CURSOR_TABLE into hl, adding
W_CURSOR_POS and then writing to the address stored there.