Implement RNG
Create a subroutine RNGWord, that returns a random word of data. Using a
word as a 6-sided dice roll can be represented with 3 bits, so 5 would
need 15 bits, or ~2 bytes. At the moment, the function is seeded from
the random-ish data in uninitialised WRAM, and then whenever more is
needed, is xor-ed with the current clock value. In future, the seed
value should probably be changed to be with clock time on button presses.