Implement Dice Variables
Create several variables for holding the value of dice in different
formats, so the different score methods can be calculated fairly easily.
At the moment, dice rolls are fixed, and in future they should be
replaced with the rng functions.
InitDice is a function that initialises the WRAM used by all the
variables, and UpdateDice should be called on each dice roll to update
the scores, which is handled by the RollDice function.