Add score calculation functions
Add a function to load into register a the score for a given
combination. Some take parameters specifying which type the score is
for, others don't. Also correct some of the documentation on
DICE_TYPE_SUM and _COUNT as the ordering was wrong. Some of these
functions can probably be combined in future or streamlined a little
(for instance, Chance can probably be reduced to a macro and OfAKind and
Yatzy combined somehow).