Update scores on screen after dice roll
The scores for all the various categories are now stored in WRAM during
the RollDice function. Then, after waiting for a VBlank to occur,
they're written to VRAM by the drawScores function. This separation is
needed as calculating scores requires more cycles than there are
available in a single VBlank cycle.

Values for the scores are stored in BCD format, having been converted by
Damian Yerrick's bcd.z80 library. All credit and thanks go to them for
producing that file - no changes have been made other than to rename it
to bcd.asm to fit in with the rest of the project.