Implement random dice rolling
Uses rng provided by rand.asm, which is a renamed rand.z80 for which all
credit and thanks go to Damian Yerrick. At the moment, a modulus-ish
command (1 is added so the range is 1 to n) is used to coerce a bit into
a dice roll - this does not produce a uniform probability distribution,
so should probably be changed in future, but is good enough for now.