Add basic input and some dice tiles
Using a new function ReadJoypad, credit and thanks to bitnenfer, read
the joypad every frame. If the A button has just been pressed, roll the
dice, calculate scores, then draw the scores, and some new tiles added
this commit, to the screen. The tiles are stored in the unused ASCII
control char section of the tile data.

One thing to note is that the Pan Docs recommend testing buttons for a
few times to allow values to settle, and others have mentioned button
bounce. I've not checked this on real hardware, but there could be
issues.