Create held cursor
Create a new cursor, represented by the ^ character, that appears below
the held row and will be used to hold and un-hold dice. It's position is
stored in the W_HELD_POS variable, which has it's last bit (7th)
corresponding to whether or not the hold cursor is currently visible.

The UpdateHoldCursor function, similarly to the UpdateCursor function,
manages the value store in W_HELD_POS and where the ^ character is drawn
on screen. It should only be called during a VBlank period.