Replace MoveCursor with a more generic GameDPAD
The old MoveCursor function required the main loop to manually take the
value stored in a after it returned and pass it to UpdateCursor, which
was messy. Now, it has been replaced by GameDPAD, which encapsulates all
the various game-related function calls and memory updates, so the main
loop is only required to pass in the new DPAD values. In addition to
making the program more contained, it also means UpdateCursor now has no
need to be globally exported and is only called when its position
changed.

Furthermore, the new GameDPAD function handles left/right movement in
the score section, as well as the visibility of the hold cursor
(determined from cursor position) and its position (left/right DPAD
presses when visible), calling UpdateHeldCursor if it changed.