Implement held dice functionality
Create a new variable in dice.asm, DICE_HELD, that is a bit array of the
currently held dice. When RollDice is called, a new value is only
generated for the dice that aren't held, otherwise the existing value is
used (note that, when there is no dice value present, nothing occurs).

In game.asm, a function, DrawHeld, has been created that draws a *
symbol beneath all the held die, and GameAction, called whenever the A
button is pressed, has been modified, adding a section where the dice
under the held cursor has it's held toggled if the held cursor is
visible.