Rearrange locations of score variables in memory
So that retrieving the score the cursor is next to, change the order of
the score variables in memory so that their order corresponds with the
order of the buttons on-screen and cursor position values.
As a result, the score of the button next to the cursor is:
score = [W_DICE_SCORES + (W_CURSOR_POS - 3)]
where [x] is the value stored in memory address x.