Implement cursor locking
If the last bit of W_CURSOR_POS is set, change the GameDPAD function to
exit immediately and ignore the input. This serves to lock the cursor in
place until it is unlocked, which so far is only done by the GameAction
function, to force the player to click a specific button.
Also, lock the cursor during the SetupGame function. This way, the
player has to roll the dice before they can score or hold a die.