Reverse input check order
Swap the input checking in the main game loop so that a DPAD game change
is checked for before checking for button inputs. This is so that if, in
the same frame, both the A button and the DPAD is pressed, only the DPAD
input is evaluated and the A button input ignored. This makes more sense
in my head than the other way round, even though a frame-perfect input
of both button and DPAD is unlikely.