Add scoring system
Create a new variable W_SCORE to hold the current game's score. It's
updated during the GameAction function whenever the cursor isn't next to
the roll or held buttons. Therefore, whenever GameAction is called, an
assumption is made that the cursor position is in range (if it isn't
errors will occur).

Then the score is obtained by adding the cursor pos to the location of
the scores in memory (see commit 4736180c096842e16ca878627cb45ad4579f28ec for
more info). As the score is spaced over two bytes, increment the second
if the carry bit is set after adding the new value to the first byte -
do not worry about overflow in the second byte as the highest possible
score in a game of yatzy is 374.

A new function, DrawHighscores, is then called, which draws the
score, which is stored in 2 bytes formatted as BCD, onto the screen.