Track which score categories have been used
Create a new 2 byte bit array, W_USED_SCORES, which tracks which scoring
categories have been used in the current game. When scoring is
attempted, W_USED_SCORES is checked to see if the category has already
been used and if not, is set.

UpdateCursor has been updated so that when the cursor moves away from a
category that has been used, a '*' is drawn in its place. This means
that which categories have been used will be visible as long as the
cursor is on them when their status changes.

Therefore, when used scores is reset after a game, it will be necessary
to manually erase each of the "*" symbols.