Add a dice roll count
Add a new variable W_DICE, which keeps track of the number of dice rolls
remaining. The player can only roll the dice when the number is above 80
and when they do, the value is decremented. Whenever a player chooses a
scoring category, W_DICE is reset back to 3, the dice and which dice are
held are reset, and the player cursor is moved back to be next to the
roll button and is locked.