Refactor DrawScores, freezing scored categories
Only draw scores to the screen for categories that haven't already been
scored (ie. have a 0 in W_USED_SCORES variable). This has the effect of
'freezing' the value that a category was scored for. In the process,
refactor the DrawScores function to make use of loops. As scores are now
organised in memory in cursor pos order (which is also the order they're
drawn in), we can use a register to keep track of where to read the
score from next.

Note that this doesn't save cycles (doesn't lose them either), it just
makes it easier to reformat the function to be more loop oriented.