Add bonus category
The bonus category gives 50 points to the player when the sum of all
scored single categories is greater than or equal to 63.

Adds a W_SINGLE_SUM variable which tracks the sum of all the singles
categories as well as whether the bonus has already been scored. The
values are update during the GameAction function and drawn to the screen
during the DrawHighscores function.