Create GameOver function
Checks whether the game is over yet by comparing the used scoring
categories variable to the value it holds when all categories have been
scored. If so, the main function jumps back and sets up the game again.

In future, this should be replaced by displaying a game over message,
and asking the player if they wish to play again. Also, the function is
currently called once every frame - this should be changed to check only
when a game action occurred.