Create game over screen
After game is complete, display game over message and wait for the
player to either press a to play again or b to goto menu.

The menu isn't implemented yet, so pressing b will hang.

When replaying, blank characters are drawn over the game over message so
none of it shows through once the game begins.