Add screen parameter to AT macro
So tiles can be drawn to either the first or second screen. All existing
AT uses stay the same, drawing to the first screen, but this will be
used in future for drawing the menu.

Drawing the menu to the second screen saves having to redraw/clear the
screen every time a game starts/ends.