Change position of code in final ROM
Fill up the interrupt vectors section so that no code is placed between
it and the entry point section. Doesn't make any difference, but I like
the way it keeps the final ROM 'clean'.

Also, force the `Main` section to be placed immediately after where the
header will be, so that the entry point can use a relative jump to get
there, rather than a slower `jp` command which it had to beforehand.