Create game menu
Menu displays a pair of dice, the title of the game, as well as a prompt
to the player to press start - when they do the game is loaded.

All menu graphics are drawn to the second screen. This is so that when
switching to the game, there is no need to erase what remains of the
menu before drawing the game text. In future, this could be optimised so
that the game and menu screens are both only drawn once, and then
switched between whenever needed. Drawing to the second screen uses the
new parameter in the AT macro.

As an additional time save, the menu tiles are loaded into the second
and third bank. This is so they don't overlap with the game tiles and
therefore only need loading on boot. As a result, the menu.inc tile map
needed to be manually modified from what was output by gbtdg to refer to
the correct tiles when in $8800 addressing mode.

Input works in a very similar way to during the Game, falling through to
it when the start button is pressed.