Seed RNG using a random value
Whenever the game is loaded, seed the RNG using the value in the DIV
register (ie. the clock register). This will occur even if the game is
restarted after the player presses "play again" on the game over menu.
This should provide some variety in the dice rolls, though I doubt it's
massively secure as the seeding will always occur a certain number of
cycles after a vblank. However this should matter too much as this is
just a game.

One point to note is that the RNG library I'm using takes a 16-bit value
for a seed and I'm only providing a 8-bit value.

For debugging purposes, it is still possible to set a fixed seed. If the
Makefile variable RNG has a value other than -1, that value will be
fixed in the ROM and all games will start with the RNG seeded to that
value. If the provided value is out of range for a 16-bit integer, it
will be truncated.