Fix bonus not being scored when it should
The original (rather awful) code would score the bonus when the tens
digit of the single's sum was greater than 6 and the units digit greater
than 3. This means that if the sum of all the single's categories totals
to to a number greater than 63 with a units digit less than 3 (eg. 71,
82, etc...), the bonus won't be scored.

I've also updated the incorrect reasoning behind not checking for the
hundreds column - my assumption was correct, but I was wrong about why.