Reduce the number of times the screen is disabled
Before this commit, the screen was disabled when changing between the
menu and the game to allow for redrawing _all_ game text to the screen.
This was obviously inefficient, in addition to causing a brief flash of
white, as the game text doesn't change between each game.

I've refactored the code to move most of the drawing of game and menu
tiles to when the game is first booted such that disabling the screen
is no longer necessary after then. However, some text (including scores,
bonuses and text overwritten by the game over message) still needs to be
drawn at the beginning of each game, which now occurs during the game
setup function.

A large change is present in the game text. As the game text is now only
drawn at the start of the game, the bodge where a preceding space in the
text overwrote the used category marker no longer works. Instead, a
simple repeated instruction has been placed in the SetupGame function
that goes through the cursor positions list and writes a zero to every
address.

One thing to note is some of the new labels (CURSOR_TABLE_CATEGORIES,
LABEL_TEXT_ROLLHELD) - none of them mark new data, but new/changed
positions in the data to allow for the fact the drawing of text has been
split in some places.