Add animated transition between menu and game
Create a new function Transition in the Main section that animates the
Game screen moving to cover or reveal the menu, based on which is
currently being viewed. It uses STAT interrupts to switch between
drawing the screen with the menu on and the screen with the game on at
some point in the middle of the frame. This is called when going from
the menu to the game and when going from the game over screen back to
the menu.

There is a limited amount of time between a STAT interrupt being
triggered and the next line being drawn, so I've stored most of the
constants used into registers and pre-calculated some of the needed
values (such as what to set rSCY to) in order to save cycles following
the STAT interrupt. Without this, graphical glitches appeared in the top
left of the game screen during the transition.

I've also made some changes to GameSetup and the way it is called.
When Transition is called, all the text and values needed for a new game
(eg. clearing the category used markers and drawing the "press a"
prompt) have to have been already drawn, otherwise old or non-existent
will be visible. As such, GameSetup is now called during setup (to ready
the first game) and, somewhat counter-intuitively, _after_ a game is
finished.

Furthermore, GameSetup now clears all the old category scores, as they
were being left over from the previous game until the first roll. I
chose to manually draw characters over just using the already defined
DrawScores function as I found that this would take longer than a single
VBlank, causing graphical issues. A similar change is made with the
drawing of the cursor - a call to UpdateCursor is no longer made - in
order to save cycles.