Add Gameboy Color support
On a gameboy color or advance, running a gameboy-only game has a really
ugly palette, so this commit makes the built ROM color-compatible, then
uses color-only function to change the main palette. This way, it still
works on an original gameboy, and has a better-looking green palette on
GBC/GBA.

In future, I might rework this so that the menu is one palette (so the
dice can be a separate color from the text), and the game another, but
this would require changing it during the Transtion function.