Inline drawing of Game Over to save cycles
A call to GameAction and subsequently to DrawGameOver used to take more
cycles than there are in a single VBlank, leading to there being a
single frame where the "Press A" prompt is visible.

I found that annoying, so by inlining the drawing of the game over
message I've shortened the runtime of DrawGameOver by about 70%,
allowing processing the game action and drawing the game over message to
all occur in a single VBlank.