Add a highscore system
Create a new variable W_HIGHSCORE, that stores the highest score
achieved by the player in a game of yatzy. It is updated whenever the
user scores a category that takes their game total over the highscore.

Whenever the high score increases, it is saved to battery-backed
external RAM so that it persists after the console is powered off. This
is then read on boot to restore the old value.

To facilitate this, there are some new functions. ReadHighscore and
WriteHighscore read and write the highscore to external RAM, with
additional checks in ReadHighscore to ensure that the highscore is
valid. In the event that it isn't (for instance, when no saved value
exists), a highscore of 0 is set to W_HIGHSCORE and saved to external
RAM.

My aim was to make highscore saving compatible with as many MBCs as
possible, so in external RAM, the highscore is written with each byte
holding a single nibble of the highscore in the low half of the byte.
For example, a highscore of 163 would be represented as follows:

In work RAM (W_HIGHSCORE):

..., $63, $01, ...

In external RAM (noting that the highscore is always saved to the start
of the external RAM):

$06, $03, $00, $01, ...

Additionally, there is a new function DrawHighscore that draws the
W_HIGHSCORE variable to the screen below the game score. This is called
whenever the highscore increases and once at boot. To avoid confusion
with the existing DrawHighscores (which never drew the highscore),
DrawHighscores is now called DrawCurrentScores.

A change to the Makefile has also been made to specify the MBC type and
RAM amount needed, though an effort has been made to make most MBCs and
RAM sizes compatible so that most flash carts can be used.